% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{remove_axis}
\alias{remove_axis}
\title{Remove an axis from a ggplot}
\usage{
remove_axis(axis = "x")
}
\arguments{
\item{axis}{Which axis? \code{"x"}, \code{"y"}, or \code{"both"}. Default \code{"x"}.}
}
\value{
A \code{ggplot2::theme()} call.
}
\description{
Equivalent to \code{urbnthemes::remove_axis()}.
}
