% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_vvn}
\alias{scale_color_vvn}
\alias{scale_colour_vvn}
\alias{scale_fill_vvn}
\title{Discrete color scale using VVN brand palettes}
\usage{
scale_color_vvn(palette = "main", reverse = FALSE, alpha = 1, ...)

scale_colour_vvn(palette = "main", reverse = FALSE, alpha = 1, ...)

scale_fill_vvn(palette = "main", reverse = FALSE, alpha = 1, ...)
}
\arguments{
\item{palette}{VVN palette name (see \code{\link[=vvn_palette]{vvn_palette()}}). Default \code{"main"}.}

\item{reverse}{Reverse palette order? Default \code{FALSE}.}

\item{alpha}{Opacity (0–1). Default \code{1}.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Drop-in replacement for \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_color_discrete()}}.
}
\examples{
library(ggplot2)
ggplot(ggplot2::mpg, aes(class, fill = class)) +
  geom_bar() + scale_fill_vvn()
}
