% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_vvn_c}
\alias{scale_color_vvn_c}
\alias{scale_colour_vvn_c}
\alias{scale_fill_vvn_c}
\title{Continuous color scale using VVN sequential / diverging palettes}
\usage{
scale_color_vvn_c(palette = "maroon_seq", reverse = FALSE, ...)

scale_colour_vvn_c(palette = "maroon_seq", reverse = FALSE, ...)

scale_fill_vvn_c(palette = "maroon_seq", reverse = FALSE, ...)
}
\arguments{
\item{palette}{One of \code{"maroon_seq"}, \code{"orange_seq"}, \code{"navy_seq"},
\code{"diverging"}, \code{"gray_seq"}. Default \code{"maroon_seq"}.}

\item{reverse}{Reverse palette? Default \code{FALSE}.}

\item{...}{Passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}.}
}
\description{
Continuous color scale using VVN sequential / diverging palettes
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() + scale_fill_vvn_c("maroon_seq")
}
