% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vvn.R
\name{undo_vvn_defaults}
\alias{undo_vvn_defaults}
\title{Revert to ggplot2 defaults}
\usage{
undo_vvn_defaults()
}
\value{
Invisibly \code{NULL}.
}
\description{
Undoes \code{set_vvn_defaults()} and restores ggplot2 factory settings.
}
