% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvn_shiny.R
\name{vvn_button}
\alias{vvn_button}
\title{VVN action button}
\usage{
vvn_button(
  inputId,
  label,
  icon = NULL,
  style = c("primary", "secondary", "outline"),
  width = NULL
)
}
\arguments{
\item{inputId}{Shiny input ID.}

\item{label}{Button label.}

\item{icon}{Optional shiny icon. Default \code{NULL}.}

\item{style}{Button style variant. Default \code{"primary"}.}

\item{width}{CSS width. Default \code{NULL}.}
}
\value{
A Shiny \code{actionButton}.
}
\description{
A styled \code{actionButton}. Style options: \code{"primary"} (maroon), \code{"secondary"}
(orange), \code{"outline"} (bordered maroon).
}
