% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvn_shiny.R
\name{vvn_filter}
\alias{vvn_filter}
\title{VVN-branded select filter}
\usage{
vvn_filter(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  width = "100\%"
)
}
\arguments{
\item{inputId}{Shiny input ID.}

\item{label}{Display label.}

\item{choices}{Named vector/list of choices.}

\item{selected}{Default value(s). Default \code{NULL}.}

\item{multiple}{Allow multiple selections? Default \code{FALSE}.}

\item{width}{CSS width. Default \code{"100\%"}.}
}
\value{
A Shiny \code{tagList}.
}
\description{
A Shiny \code{selectInput} wrapped in a VVN-styled container.
}
