% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{vvn_note}
\alias{vvn_note}
\title{Add a VVN note / annotation line}
\usage{
vvn_note(text)
}
\arguments{
\item{text}{Note text.}
}
\value{
A \code{ggplot2::labs()} call.
}
\description{
Equivalent to \code{urbnthemes::urbn_note()}. Adds a footnote-style annotation
below the plot using \code{ggplot2::labs(caption = ...)}.
}
